fix(shared): mark rate-limit timestamps as integer
resets_at and observed_at are epoch-based timestamps; the rest of
protocol.ts uses z.number().int() consistently for the same kind of
field (expiresAt, quietUntil, seq, etc.). Tightens validation so a
malformed payload with floats can't sneak through.

+++ packages/shared/src/protocol.ts
@@ -254,13 +254,13 @@ export const SessionChildChangedSchema = z.object({
 
 export const RateLimitWindowSchema = z.object({
   used_percentage: z.number(),
-  resets_at:       z.number(),
+  resets_at:       z.number().int(),  // Unix seconds (matches CC's source)
 });
 
 export const RateLimitsStateSchema = z.object({
   five_hour:    RateLimitWindowSchema.nullable(),
   seven_day:    RateLimitWindowSchema.nullable(),
-  observed_at:  z.number(),
+  observed_at:  z.number().int(),     // Unix milliseconds (hub-stamped via Date.now())
 });
 
 export const SessionRateLimitsSchema = z.object({
